more proper bitwise up handling and casting

+++ src/main/object/sm_space.c
@@ -4,7 +4,7 @@
 
 // General purpose space of memory
 sm_space *sm_new_space(uint32_t size) {
-  sm_space *new_space = sm_malloc(sizeof(sm_space) + sm_round_size64(size));
+  sm_space *new_space = sm_malloc((uint32_t)sizeof(sm_space) + sm_round_size64(size));
   new_space->my_type  = SM_SPACE_TYPE;
   new_space->size     = size;
   return new_space;
@@ -12,7 +12,7 @@ sm_space *sm_new_space(uint32_t size) {
 
 //
 sm_space *sm_new_space_at(sm_heap *h, uint32_t size) {
-  sm_space *new_space = sm_malloc_at(h, sizeof(sm_space) + sm_round_size64(size));
+  sm_space *new_space = sm_malloc_at(h, (uint32_t)sizeof(sm_space) + sm_round_size64(size));
   new_space->my_type  = SM_SPACE_TYPE;
   new_space->size     = size;
   return new_space;
